First steps into a UI agnostic input crate
