+++ crates/gfx/Cargo.toml
@@ -13,4 +13,5 @@ css = { path = "../css" }
 layout = { path = "../layout" }
 core_types = { path = "../core_types" }
 html = { path = "../html" }
+input_core = { path = "../input_core" }
 url = { workspace = true }
