+++ crates/gfx/src/input/store.rs
@@ -1,387 +1,253 @@
-use crate::util::clamp_to_char_boundary;
+//! Input value store wrapper that bridges `html::Id` to `input_core::InputId`.
+//!
+//! This module provides a thin wrapper around `input_core::InputValueStore` that
+//! accepts `html::Id` and automatically converts it to the UI-agnostic `InputId`.
+
 use html::Id;
-use std::borrow::Cow;
-use std::collections::HashMap;
+use input_core::{InputId, InputValueStore as CoreInputValueStore};
 
-fn prev_cursor_boundary(s: &str, i: usize) -> usize {
-    let i = clamp_to_char_boundary(s, i);
-    if i == 0 {
-        return 0;
-    }
-    s[..i]
-        .char_indices()
-        .last()
-        .map(|(idx, _)| idx)
-        .unwrap_or(0)
-}
+// Re-export SelectionRange directly since it has no Id dependency
+pub use input_core::SelectionRange;
 
-fn next_cursor_boundary(s: &str, i: usize) -> usize {
-    let i = clamp_to_char_boundary(s, i);
-    if i >= s.len() {
-        return s.len();
-    }
-
-    let mut it = s[i..].char_indices();
-    let _ = it.next(); // current char at 0
-    it.next().map(|(idx, _)| i + idx).unwrap_or(s.len())
-}
-
-#[derive(Clone, Debug)]
-struct InputState {
-    value: String,
-    value_rev: u64,
-    checked: bool,
-    /// Caret position as a byte index into `value` (always on a UTF-8 char boundary).
-    caret: usize,
-    /// Selection anchor as a byte index into `value` (UTF-8 char boundary).
-    ///
-    /// When `Some(anchor)`, the selection range is `min(anchor, caret)..max(anchor, caret)`.
-    selection_anchor: Option<usize>,
-    /// Horizontal scroll offset in px for single-line inputs.
-    scroll_x: f32,
-    /// Vertical scroll offset in px for multi-line text controls (e.g. `<textarea>`).
-    scroll_y: f32,
-    cursor_boundaries: Vec<usize>,
-    cursor_boundaries_dirty: bool,
-}
-
-impl Default for InputState {
-    fn default() -> Self {
-        Self {
-            value: String::new(),
-            value_rev: 0,
-            checked: false,
-            caret: 0,
-            selection_anchor: None,
-            scroll_x: 0.0,
-            scroll_y: 0.0,
-            cursor_boundaries: Vec::new(),
-            cursor_boundaries_dirty: true,
-        }
-    }
-}
-
+/// Wrapper around `input_core::InputValueStore` that uses `html::Id`.
+///
+/// This provides the same API as the core store but accepts `html::Id` directly,
+/// making it seamless to use with the DOM-based browser engine.
 #[derive(Clone, Debug, Default)]
 pub struct InputValueStore {
-    values: HashMap<Id, InputState>,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct SelectionRange {
-    pub start: usize,
-    pub end: usize,
+    inner: CoreInputValueStore,
 }
 
 impl InputValueStore {
+    /// Create a new, empty input value store.
+    #[inline]
     pub fn new() -> Self {
         Self {
-            values: HashMap::new(),
+            inner: CoreInputValueStore::new(),
         }
     }
 
+    /// Returns `true` if an entry exists for this input.
+    #[inline]
     pub fn has(&self, id: Id) -> bool {
-        self.values.contains_key(&id)
+        self.inner.has(to_input_id(id))
     }
 
+    /// Get the full state tuple for an input.
+    ///
+    /// Returns `(value, caret, selection, scroll_x, scroll_y)` if the input exists.
+    #[inline]
     pub fn get_state(&self, id: Id) -> Option<(&str, usize, Option<SelectionRange>, f32, f32)> {
-        self.values.get(&id).map(|s| {
-            let sel = selection_range(&s.value, s.selection_anchor, s.caret);
-            (s.value.as_str(), s.caret, sel, s.scroll_x, s.scroll_y)
-        })
+        self.inner.get_state(to_input_id(id))
     }
 
-    /// Monotonic revision counter for the input's value. Increments on any text change.
+    /// Monotonic revision counter for the input's value.
+    ///
+    /// Increments on any text change. Useful for cache invalidation.
+    #[inline]
     pub fn value_revision(&self, id: Id) -> u64 {
-        self.values.get(&id).map(|s| s.value_rev).unwrap_or(0)
+        self.inner.value_revision(to_input_id(id))
     }
 
-    /// Returns the stored value for this key, if any.
+    /// Returns the stored value for this input, if any.
+    #[inline]
     pub fn get(&self, id: Id) -> Option<&str> {
-        self.values.get(&id).map(|s| s.value.as_str())
+        self.inner.get(to_input_id(id))
     }
 
     /// Returns the current caret byte index for this input, if any.
+    #[inline]
     pub fn caret(&self, id: Id) -> Option<usize> {
-        self.values.get(&id).map(|s| s.caret)
+        self.inner.caret(to_input_id(id))
     }
 
+    /// Returns `true` if this checkbox/radio input is checked.
+    #[inline]
     pub fn is_checked(&self, id: Id) -> bool {
-        self.values.get(&id).is_some_and(|s| s.checked)
+        self.inner.is_checked(to_input_id(id))
     }
 
+    /// Set the checked state for a checkbox/radio input.
+    ///
+    /// Returns `true` if the state actually changed.
+    #[inline]
     pub fn set_checked(&mut self, id: Id, checked: bool) -> bool {
-        let st = self.values.entry(id).or_default();
-        let changed = st.checked != checked;
-        st.checked = checked;
-        changed
+        self.inner.set_checked(to_input_id(id), checked)
     }
 
+    /// Toggle the checked state for a checkbox/radio input.
+    ///
+    /// Returns `true` if the state changed (which is always true for toggle).
+    #[inline]
     pub fn toggle_checked(&mut self, id: Id) -> bool {
-        let st = self.values.entry(id).or_default();
-        let new_val = !st.checked;
-        let changed = st.checked != new_val;
-        st.checked = new_val;
-        changed
+        self.inner.toggle_checked(to_input_id(id))
     }
 
+    /// Ensure an entry exists with the initial checked state.
+    ///
+    /// If an entry already exists, this is a no-op.
+    #[inline]
     pub fn ensure_initial_checked(&mut self, id: Id, initial_checked: bool) {
-        self.values.entry(id).or_insert(InputState {
-            checked: initial_checked,
-            ..InputState::default()
-        });
+        self.inner
+            .ensure_initial_checked(to_input_id(id), initial_checked)
     }
 
-    /// Set/overwrite the value for this input key.
+    /// Set/overwrite the value for this input.
+    ///
+    /// This resets the caret to the end and clears any selection.
+    #[inline]
     pub fn set(&mut self, id: Id, value: String) {
-        let caret = clamp_to_char_boundary(&value, value.len());
-        let checked = self.values.get(&id).is_some_and(|s| s.checked);
-        let value_rev = self
-            .values
-            .get(&id)
-            .map(|s| s.value_rev.wrapping_add(1))
-            .unwrap_or(0);
-        self.values.insert(
-            id,
-            InputState {
-                value,
-                value_rev,
-                checked,
-                caret,
-                selection_anchor: None,
-                scroll_x: 0.0,
-                scroll_y: 0.0,
-                cursor_boundaries: Vec::new(),
-                cursor_boundaries_dirty: true,
-            },
-        );
+        self.inner.set(to_input_id(id), value)
     }
 
-    /// Ensure a key exists; if missing, inserts the provided initial value.
+    /// Ensure an entry exists; if missing, inserts the provided initial value.
+    #[inline]
     pub fn ensure_initial(&mut self, id: Id, initial: String) {
-        let caret = clamp_to_char_boundary(&initial, initial.len());
-        self.values.entry(id).or_insert(InputState {
-            value: initial,
-            value_rev: 0,
-            checked: false,
-            caret,
-            selection_anchor: None,
-            scroll_x: 0.0,
-            scroll_y: 0.0,
-            cursor_boundaries: Vec::new(),
-            cursor_boundaries_dirty: true,
-        });
+        self.inner.ensure_initial(to_input_id(id), initial)
     }
 
-    /// When an input is focused, clamp caret to a valid UTF-8 boundary and clear selection.
+    /// Called when an input gains focus.
+    ///
+    /// Clamps caret to a valid UTF-8 boundary and clears selection.
+    #[inline]
     pub fn focus(&mut self, id: Id) {
-        if let Some(st) = self.values.get_mut(&id) {
-            clamp_state(st);
-            clear_selection(st);
-        }
+        self.inner.focus(to_input_id(id))
     }
 
+    /// Called when an input loses focus.
+    ///
+    /// Clamps caret to a valid boundary and clears selection.
+    #[inline]
     pub fn blur(&mut self, id: Id) {
-        if let Some(st) = self.values.get_mut(&id) {
-            clamp_state(st);
-            clear_selection(st);
-        }
+        self.inner.blur(to_input_id(id))
     }
 
+    /// Clear all stored input state.
+    ///
+    /// Typically called on navigation to reset document state.
+    #[inline]
     pub fn clear(&mut self) {
-        self.values.clear();
+        self.inner.clear()
     }
 
+    /// Insert text at the current caret position (single-line mode).
+    ///
+    /// Newlines are stripped. If there is a selection, it is replaced.
+    #[inline]
     pub fn insert_text(&mut self, id: Id, s: &str) {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-        let s = filter_single_line(s);
-        if s.is_empty() {
-            return;
-        }
-
-        delete_selection_if_any(st);
-
-        let caret = clamp_to_char_boundary(&st.value, st.caret);
-        st.value.insert_str(caret, &s);
-        st.caret = caret + s.len();
-        st.caret = clamp_to_char_boundary(&st.value, st.caret);
-        mark_text_dirty(st);
+        self.inner.insert_text(to_input_id(id), s)
     }
 
+    /// Insert text at the current caret position (multi-line mode).
+    ///
+    /// Newlines are normalized (CRLF/CR → LF). If there is a selection, it is replaced.
+    #[inline]
     pub fn insert_text_multiline(&mut self, id: Id, s: &str) {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-
-        let s = normalize_newlines(s);
-        if s.is_empty() {
-            return;
-        }
-
-        delete_selection_if_any(st);
-
-        let caret = clamp_to_char_boundary(&st.value, st.caret);
-        st.value.insert_str(caret, &s);
-        st.caret = caret + s.len();
-        st.caret = clamp_to_char_boundary(&st.value, st.caret);
-        mark_text_dirty(st);
+        self.inner.insert_text_multiline(to_input_id(id), s)
     }
 
+    /// Delete the character before the caret (backspace).
+    ///
+    /// If there is a selection, deletes the selection instead.
+    #[inline]
     pub fn backspace(&mut self, id: Id) {
-        if let Some(st) = self.values.get_mut(&id) {
-            clamp_state(st);
-            if delete_selection_if_any(st) {
-                return;
-            }
-
-            let caret = clamp_to_char_boundary(&st.value, st.caret);
-            if caret == 0 {
-                return;
-            }
-
-            let prev = prev_cursor_boundary(&st.value, caret);
-            st.value.drain(prev..caret);
-            st.caret = clamp_to_char_boundary(&st.value, prev);
-            mark_text_dirty(st);
-        }
+        self.inner.backspace(to_input_id(id))
     }
 
+    /// Delete the character after the caret (delete key).
+    ///
+    /// If there is a selection, deletes the selection instead.
+    #[inline]
     pub fn delete(&mut self, id: Id) {
-        if let Some(st) = self.values.get_mut(&id) {
-            clamp_state(st);
-            if delete_selection_if_any(st) {
-                return;
-            }
-
-            let caret = clamp_to_char_boundary(&st.value, st.caret);
-            if caret >= st.value.len() {
-                return;
-            }
-
-            let next = next_cursor_boundary(&st.value, caret);
-            st.value.drain(caret..next);
-            st.caret = clamp_to_char_boundary(&st.value, caret);
-            mark_text_dirty(st);
-        }
+        self.inner.delete(to_input_id(id))
     }
 
+    /// Move the caret left by one character.
+    ///
+    /// If `selecting` is true, extends/modifies the selection.
+    #[inline]
     pub fn move_caret_left(&mut self, id: Id, selecting: bool) {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-
-        if selecting {
-            if st.selection_anchor.is_none() {
-                st.selection_anchor = Some(st.caret);
-            }
-            st.caret = prev_cursor_boundary(&st.value, st.caret);
-            normalize_selection_anchor(st);
-            return;
-        }
-
-        if let Some(sel) = selection_range(&st.value, st.selection_anchor, st.caret) {
-            st.caret = sel.start;
-        } else {
-            st.caret = prev_cursor_boundary(&st.value, st.caret);
-        }
-        clear_selection(st);
+        self.inner.move_caret_left(to_input_id(id), selecting)
     }
 
+    /// Move the caret right by one character.
+    ///
+    /// If `selecting` is true, extends/modifies the selection.
+    #[inline]
     pub fn move_caret_right(&mut self, id: Id, selecting: bool) {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-
-        if selecting {
-            if st.selection_anchor.is_none() {
-                st.selection_anchor = Some(st.caret);
-            }
-            st.caret = next_cursor_boundary(&st.value, st.caret);
-            normalize_selection_anchor(st);
-            return;
-        }
-
-        if let Some(sel) = selection_range(&st.value, st.selection_anchor, st.caret) {
-            st.caret = sel.end;
-        } else {
-            st.caret = next_cursor_boundary(&st.value, st.caret);
-        }
-        clear_selection(st);
+        self.inner.move_caret_right(to_input_id(id), selecting)
     }
 
+    /// Move the caret to the start of the text.
+    ///
+    /// If `selecting` is true, extends/modifies the selection.
+    #[inline]
     pub fn move_caret_to_start(&mut self, id: Id, selecting: bool) {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-
-        if selecting {
-            if st.selection_anchor.is_none() {
-                st.selection_anchor = Some(st.caret);
-            }
-            st.caret = 0;
-            normalize_selection_anchor(st);
-        } else {
-            st.caret = 0;
-            clear_selection(st);
-        }
+        self.inner.move_caret_to_start(to_input_id(id), selecting)
     }
 
+    /// Move the caret to the end of the text.
+    ///
+    /// If `selecting` is true, extends/modifies the selection.
+    #[inline]
     pub fn move_caret_to_end(&mut self, id: Id, selecting: bool) {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-
-        if selecting {
-            if st.selection_anchor.is_none() {
-                st.selection_anchor = Some(st.caret);
-            }
-            st.caret = st.value.len();
-            normalize_selection_anchor(st);
-        } else {
-            st.caret = st.value.len();
-            clear_selection(st);
-        }
+        self.inner.move_caret_to_end(to_input_id(id), selecting)
     }
 
+    /// Select all text in the input.
+    #[inline]
     pub fn select_all(&mut self, id: Id) {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-        st.caret = st.value.len();
-        st.selection_anchor = Some(0);
-        normalize_selection_anchor(st);
+        self.inner.select_all(to_input_id(id))
     }
 
+    /// Set the caret to a specific byte position.
+    ///
+    /// If `selecting` is true, extends/modifies the selection.
+    #[inline]
     pub fn set_caret(&mut self, id: Id, caret: usize, selecting: bool) {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-
-        let caret = clamp_to_char_boundary(&st.value, caret);
-
-        set_caret_in_state(st, caret, selecting);
+        self.inner.set_caret(to_input_id(id), caret, selecting)
     }
 
+    /// Set the caret based on a viewport x-coordinate.
+    ///
+    /// Uses the provided measurement function to determine which character
+    /// boundary is closest to the given x position.
+    ///
+    /// # Arguments
+    ///
+    /// * `id` - The input ID
+    /// * `x_in_viewport` - X position relative to the visible viewport
+    /// * `selecting` - Whether to extend the selection
+    /// * `measure_prefix` - Function that measures the width of a text prefix
+    ///
+    /// # Returns
+    ///
+    /// The byte index of the new caret position.
+    #[inline]
     pub fn set_caret_from_viewport_x(
         &mut self,
         id: Id,
         x_in_viewport: f32,
         selecting: bool,
-        mut measure_prefix: impl FnMut(&str) -> f32,
+        measure_prefix: impl FnMut(&str) -> f32,
     ) -> usize {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-
-        let x_in_viewport = x_in_viewport.max(0.0);
-        let x_in_text = x_in_viewport + st.scroll_x;
-
-        let caret = {
-            ensure_cursor_boundaries(st);
-            caret_from_x_with_boundaries(
-                &st.value,
-                &st.cursor_boundaries,
-                x_in_text,
-                &mut measure_prefix,
-            )
-        };
-
-        set_caret_in_state(st, caret, selecting);
-        caret
+        self.inner.set_caret_from_viewport_x(
+            to_input_id(id),
+            x_in_viewport,
+            selecting,
+            measure_prefix,
+        )
     }
 
+    /// Update horizontal scroll to keep the caret visible.
+    ///
+    /// # Arguments
+    ///
+    /// * `id` - The input ID
+    /// * `caret_px` - The caret's x position in text coordinates
+    /// * `text_w` - Total width of the text content
+    /// * `available_w` - Width of the visible viewport
+    #[inline]
     pub fn update_scroll_for_caret(
         &mut self,
         id: Id,
@@ -389,35 +255,20 @@ impl InputValueStore {
         text_w: f32,
         available_w: f32,
     ) {
-        let st = self.values.entry(id).or_default();
-
-        let available_w = available_w.max(0.0);
-        let text_w = text_w.max(0.0);
-        let caret_px = caret_px.clamp(0.0, text_w);
-
-        if available_w <= 0.0 || text_w <= available_w {
-            st.scroll_x = 0.0;
-            return;
-        }
-
-        let max_scroll = (text_w - available_w).max(0.0);
-        let mut scroll_x = st.scroll_x.clamp(0.0, max_scroll);
-
-        // Keep the caret visible with a small margin, but don't re-center unless needed.
-        let margin: f32 = 4.0;
-        let left_limit = margin.min(available_w);
-        let right_limit = (available_w - margin).max(left_limit);
-
-        let caret_in_view = caret_px - scroll_x;
-        if caret_in_view < left_limit {
-            scroll_x = (caret_px - left_limit).max(0.0);
-        } else if caret_in_view > right_limit {
-            scroll_x = (caret_px - right_limit).min(max_scroll);
-        }
-
-        st.scroll_x = scroll_x;
+        self.inner
+            .update_scroll_for_caret(to_input_id(id), caret_px, text_w, available_w)
     }
 
+    /// Update vertical scroll to keep the caret visible (for multi-line inputs).
+    ///
+    /// # Arguments
+    ///
+    /// * `id` - The input ID
+    /// * `caret_y` - The caret's y position in text coordinates
+    /// * `caret_h` - Height of the caret/line
+    /// * `text_h` - Total height of the text content
+    /// * `available_h` - Height of the visible viewport
+    #[inline]
     pub fn update_scroll_for_caret_y(
         &mut self,
         id: Id,
@@ -426,38 +277,27 @@ impl InputValueStore {
         text_h: f32,
         available_h: f32,
     ) {
-        let st = self.values.entry(id).or_default();
-
-        let available_h = available_h.max(0.0);
-        let text_h = text_h.max(0.0);
-        let caret_h = caret_h.max(0.0);
-        let caret_y = caret_y.clamp(0.0, text_h);
-
-        if available_h <= 0.0 || text_h <= available_h {
-            st.scroll_y = 0.0;
-            return;
-        }
-
-        let max_scroll = (text_h - available_h).max(0.0);
-        let mut scroll_y = st.scroll_y.clamp(0.0, max_scroll);
-
-        // Keep the caret visible with a small margin, but don't re-center unless needed.
-        let margin: f32 = 4.0;
-        let top_limit = margin.min(available_h);
-        let bottom_limit = (available_h - margin).max(top_limit);
-
-        let caret_top_in_view = caret_y - scroll_y;
-        let caret_bottom_in_view = caret_top_in_view + caret_h;
-
-        if caret_top_in_view < top_limit {
-            scroll_y = (caret_y - top_limit).max(0.0);
-        } else if caret_bottom_in_view > bottom_limit {
-            scroll_y = (caret_y + caret_h - bottom_limit).min(max_scroll);
-        }
-
-        st.scroll_y = scroll_y;
+        self.inner
+            .update_scroll_for_caret_y(to_input_id(id), caret_y, caret_h, text_h, available_h)
     }
 
+    /// Set the caret based on a viewport x-coordinate within a specific line range.
+    ///
+    /// Used for multi-line text where each line needs independent measurement.
+    ///
+    /// # Arguments
+    ///
+    /// * `id` - The input ID
+    /// * `x_in_viewport` - X position relative to the line's start
+    /// * `selecting` - Whether to extend the selection
+    /// * `range_start` - Start byte index of the line
+    /// * `range_end` - End byte index of the line
+    /// * `measure_range_prefix` - Function that measures the width of a line prefix
+    ///
+    /// # Returns
+    ///
+    /// The byte index of the new caret position.
+    #[inline]
     pub fn set_caret_from_viewport_x_in_range(
         &mut self,
         id: Id,
@@ -465,487 +305,21 @@ impl InputValueStore {
         selecting: bool,
         range_start: usize,
         range_end: usize,
-        mut measure_range_prefix: impl FnMut(&str) -> f32,
+        measure_range_prefix: impl FnMut(&str) -> f32,
     ) -> usize {
-        let st = self.values.entry(id).or_default();
-        clamp_state(st);
-
-        let range_start = clamp_to_char_boundary(&st.value, range_start);
-        let range_end = clamp_to_char_boundary(&st.value, range_end).max(range_start);
-        let x_in_viewport = x_in_viewport.max(0.0);
-
-        let caret = {
-            ensure_cursor_boundaries(st);
-
-            let start_idx = st.cursor_boundaries.partition_point(|&b| b < range_start);
-            let end_idx = st.cursor_boundaries.partition_point(|&b| b <= range_end);
-            let boundaries = &st.cursor_boundaries[start_idx..end_idx];
-
-            caret_from_x_with_boundaries_in_range(
-                &st.value,
-                boundaries,
-                range_start,
-                x_in_viewport,
-                &mut measure_range_prefix,
-            )
-        };
-
-        set_caret_in_state(st, caret, selecting);
-        caret
+        self.inner.set_caret_from_viewport_x_in_range(
+            to_input_id(id),
+            x_in_viewport,
+            selecting,
+            range_start,
+            range_end,
+            measure_range_prefix,
+        )
     }
 }
 
-fn selection_range(value: &str, anchor: Option<usize>, caret: usize) -> Option<SelectionRange> {
-    let anchor = anchor?;
-
-    let a = clamp_to_char_boundary(value, anchor);
-    let c = clamp_to_char_boundary(value, caret);
-    if a == c {
-        return None;
-    }
-
-    Some(SelectionRange {
-        start: a.min(c),
-        end: a.max(c),
-    })
-}
-
-fn set_caret_in_state(st: &mut InputState, caret: usize, selecting: bool) {
-    let caret = clamp_to_char_boundary(&st.value, caret);
-
-    if selecting {
-        if st.selection_anchor.is_none() {
-            st.selection_anchor = Some(st.caret);
-        }
-        st.caret = caret;
-        normalize_selection_anchor(st);
-    } else {
-        st.caret = caret;
-        clear_selection(st);
-    }
-}
-
-fn normalize_selection_anchor(st: &mut InputState) {
-    let Some(anchor) = st.selection_anchor else {
-        return;
-    };
-    let anchor = clamp_to_char_boundary(&st.value, anchor);
-    st.selection_anchor = Some(anchor);
-
-    // If selection collapsed, clear anchor to avoid "sticky" selection.
-    if anchor == st.caret {
-        st.selection_anchor = None;
-    }
-}
-
-fn delete_selection_if_any(st: &mut InputState) -> bool {
-    let Some(sel) = selection_range(&st.value, st.selection_anchor, st.caret) else {
-        st.selection_anchor = None;
-        st.caret = clamp_to_char_boundary(&st.value, st.caret);
-        return false;
-    };
-
-    st.value.drain(sel.start..sel.end);
-    st.caret = clamp_to_char_boundary(&st.value, sel.start);
-    st.selection_anchor = None;
-    mark_text_dirty(st);
-    true
-}
-
-fn clamp_state(st: &mut InputState) {
-    st.caret = clamp_to_char_boundary(&st.value, st.caret);
-    if let Some(a) = st.selection_anchor {
-        st.selection_anchor = Some(clamp_to_char_boundary(&st.value, a));
-    }
-    st.scroll_x = st.scroll_x.max(0.0);
-    st.scroll_y = st.scroll_y.max(0.0);
-}
-
-fn clear_selection(st: &mut InputState) {
-    st.selection_anchor = None;
-}
-
-fn mark_text_dirty(st: &mut InputState) {
-    st.cursor_boundaries_dirty = true;
-    st.value_rev = st.value_rev.wrapping_add(1);
-}
-
-fn ensure_cursor_boundaries(st: &mut InputState) {
-    if st.value.is_empty() {
-        st.cursor_boundaries.clear();
-        st.cursor_boundaries_dirty = false;
-        return;
-    }
-
-    if !st.cursor_boundaries_dirty && !st.cursor_boundaries.is_empty() {
-        return;
-    }
-
-    rebuild_cursor_boundaries(&st.value, &mut st.cursor_boundaries);
-    st.cursor_boundaries_dirty = false;
-}
-
-fn rebuild_cursor_boundaries(value: &str, out: &mut Vec<usize>) {
-    out.clear();
-    out.extend(value.char_indices().map(|(i, _)| i));
-
-    if out.first().copied() != Some(0) {
-        out.insert(0, 0);
-    }
-    if out.last().copied() != Some(value.len()) {
-        out.push(value.len());
-    }
-}
-
-fn filter_single_line(s: &str) -> Cow<'_, str> {
-    // Keep input single-line: drop CR/LF (fast-path if already clean).
-    if !s.contains('\n') && !s.contains('\r') {
-        return Cow::Borrowed(s);
-    }
-    Cow::Owned(s.chars().filter(|c| *c != '\n' && *c != '\r').collect())
-}
-
-fn normalize_newlines(s: &str) -> Cow<'_, str> {
-    // Normalize CRLF/CR to LF.
-    if !s.contains('\r') {
-        return Cow::Borrowed(s);
-    }
-
-    let mut out = String::with_capacity(s.len());
-    let mut it = s.chars().peekable();
-    while let Some(ch) = it.next() {
-        match ch {
-            '\r' => {
-                if it.peek() == Some(&'\n') {
-                    let _ = it.next();
-                }
-                out.push('\n');
-            }
-            _ => out.push(ch),
-        }
-    }
-    Cow::Owned(out)
-}
-
-#[cfg(test)]
-pub(crate) fn caret_from_x(
-    value: &str,
-    x: f32,
-    mut measure_prefix: impl FnMut(&str) -> f32,
-) -> usize {
-    let mut boundaries: Vec<usize> = Vec::new();
-    rebuild_cursor_boundaries(value, &mut boundaries);
-    caret_from_x_with_boundaries(value, &boundaries, x, &mut measure_prefix)
-}
-
-fn caret_from_x_with_boundaries(
-    value: &str,
-    boundaries: &[usize],
-    x: f32,
-    mut measure_prefix: impl FnMut(&str) -> f32,
-) -> usize {
-    if value.is_empty() || boundaries.is_empty() {
-        return 0;
-    }
-
-    let x = x.max(0.0);
-
-    // Binary search for the largest boundary whose prefix width <= x.
-    let mut lo = 0usize;
-    let mut hi = boundaries.len() - 1;
-
-    while lo < hi {
-        let mid = lo + (hi - lo).div_ceil(2);
-        let idx = boundaries[mid];
-        let w = measure_prefix(&value[..idx]).max(0.0);
-        if w <= x {
-            lo = mid;
-        } else {
-            hi = mid - 1;
-        }
-    }
-
-    let left_idx = boundaries[lo];
-    let left_w = measure_prefix(&value[..left_idx]).max(0.0);
-
-    // Snap to nearest boundary (not always floor), so clicks feel natural.
-    if lo + 1 < boundaries.len() {
-        let right_idx = boundaries[lo + 1];
-        let right_w = measure_prefix(&value[..right_idx]).max(0.0);
-        if x - left_w > right_w - x {
-            return right_idx;
-        }
-    }
-
-    left_idx
-}
-
-fn caret_from_x_with_boundaries_in_range(
-    value: &str,
-    boundaries: &[usize],
-    range_start: usize,
-    x: f32,
-    mut measure_range_prefix: impl FnMut(&str) -> f32,
-) -> usize {
-    if value.is_empty() || boundaries.is_empty() {
-        return range_start;
-    }
-
-    let x = x.max(0.0);
-
-    // Binary search for the largest boundary whose prefix width <= x.
-    let mut lo = 0usize;
-    let mut hi = boundaries.len() - 1;
-
-    while lo < hi {
-        let mid = lo + (hi - lo).div_ceil(2);
-        let idx = boundaries[mid];
-        let w = measure_range_prefix(&value[range_start..idx]).max(0.0);
-        if w <= x {
-            lo = mid;
-        } else {
-            hi = mid - 1;
-        }
-    }
-
-    let left_idx = boundaries[lo];
-    let left_w = measure_range_prefix(&value[range_start..left_idx]).max(0.0);
-
-    // Snap to nearest boundary (not always floor), so clicks feel natural.
-    if lo + 1 < boundaries.len() {
-        let right_idx = boundaries[lo + 1];
-        let right_w = measure_range_prefix(&value[range_start..right_idx]).max(0.0);
-        if x - left_w > right_w - x {
-            return right_idx;
-        }
-    }
-
-    left_idx
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn insert_text_keeps_caret_on_char_boundary() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        store.ensure_initial(id, String::new());
-        store.focus(id);
-
-        store.insert_text(id, "€"); // 3-byte UTF-8
-        let v = store.get(id).unwrap();
-        let caret = store.caret(id).unwrap();
-        assert_eq!(v, "€");
-        assert_eq!(caret, v.len());
-        assert!(v.is_char_boundary(caret));
-    }
-
-    #[test]
-    fn backspace_removes_a_full_unicode_scalar_value() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        store.set(id, "a€".to_string());
-        store.focus(id);
-
-        store.backspace(id);
-        assert_eq!(store.get(id), Some("a"));
-        let v = store.get(id).unwrap();
-        let caret = store.caret(id).unwrap();
-        assert_eq!(caret, v.len());
-        assert!(v.is_char_boundary(caret));
-    }
-
-    #[test]
-    fn invalid_caret_is_clamped_before_insert() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        store.set(id, "€".to_string());
-        store.values.get_mut(&id).unwrap().caret = 1; // invalid boundary
-
-        store.insert_text(id, "x");
-        assert_eq!(store.get(id), Some("x€"));
-        let v = store.get(id).unwrap();
-        let caret = store.caret(id).unwrap();
-        assert!(v.is_char_boundary(caret));
-    }
-
-    #[test]
-    fn move_caret_left_right_moves_by_unicode_scalar_value() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        store.set(id, "a€b".to_string());
-        store.focus(id);
-
-        // Caret starts at end.
-        assert_eq!(store.caret(id), Some("a€b".len()));
-
-        store.move_caret_left(id, false);
-        assert_eq!(store.caret(id), Some("a€".len()));
-
-        store.move_caret_left(id, false);
-        assert_eq!(store.caret(id), Some("a".len()));
-
-        store.move_caret_right(id, false);
-        assert_eq!(store.caret(id), Some("a€".len()));
-    }
-
-    #[test]
-    fn shift_arrow_creates_selection_and_backspace_deletes_it() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        store.set(id, "hello".to_string());
-        store.focus(id);
-
-        store.move_caret_left(id, true); // select last char
-        let (_v, _caret, sel, _scroll_x, _scroll_y) = store.get_state(id).unwrap();
-        assert_eq!(sel, Some(SelectionRange { start: 4, end: 5 }));
-
-        store.backspace(id);
-        assert_eq!(store.get(id), Some("hell"));
-        assert_eq!(store.caret(id), Some(4));
-
-        let (_v, _caret, sel, _scroll_x, _scroll_y) = store.get_state(id).unwrap();
-        assert_eq!(sel, None);
-    }
-
-    #[test]
-    fn typing_replaces_selection() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        store.set(id, "hello".to_string());
-        store.focus(id);
-        store.move_caret_left(id, true); // select "o"
-        store.insert_text(id, "X");
-
-        assert_eq!(store.get(id), Some("hellX"));
-        assert_eq!(store.caret(id), Some("hellX".len()));
-    }
-
-    #[test]
-    fn delete_removes_next_char() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        store.set(id, "abc".to_string());
-        store.focus(id);
-        store.move_caret_left(id, false); // caret before 'c'
-        assert_eq!(store.caret(id), Some(2));
-
-        store.delete(id);
-        assert_eq!(store.get(id), Some("ab"));
-        assert_eq!(store.caret(id), Some(2));
-    }
-
-    #[test]
-    fn delete_selection_wins_over_single_char_delete() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        store.set(id, "abcd".to_string());
-        store.focus(id);
-
-        store.move_caret_left(id, true); // select 'd'
-        store.move_caret_left(id, true); // select 'cd'
-        store.delete(id);
-
-        assert_eq!(store.get(id), Some("ab"));
-        assert_eq!(store.caret(id), Some(2));
-    }
-
-    #[test]
-    fn set_caret_supports_shift_extend_selection() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        store.set(id, "hello".to_string());
-        store.focus(id);
-
-        store.set_caret(id, 2, false);
-        let (_v, caret, sel, _scroll_x, _scroll_y) = store.get_state(id).unwrap();
-        assert_eq!(caret, 2);
-        assert_eq!(sel, None);
-
-        store.set_caret(id, 4, true);
-        let (_v, caret, sel, _scroll_x, _scroll_y) = store.get_state(id).unwrap();
-        assert_eq!(caret, 4);
-        assert_eq!(sel, Some(SelectionRange { start: 2, end: 4 }));
-
-        store.set_caret(id, 1, false);
-        let (_v, caret, sel, _scroll_x, _scroll_y) = store.get_state(id).unwrap();
-        assert_eq!(caret, 1);
-        assert_eq!(sel, None);
-    }
-
-    #[test]
-    fn caret_from_x_picks_nearest_boundary() {
-        let value = "hello";
-        let measure = |s: &str| s.chars().count() as f32 * 10.0;
-
-        assert_eq!(caret_from_x(value, 0.0, measure), 0);
-        assert_eq!(caret_from_x(value, 4.0, measure), 0); // closer to 0 than 10
-        assert_eq!(caret_from_x(value, 6.0, measure), 1); // closer to 10 than 0
-        assert_eq!(caret_from_x(value, 19.0, measure), 2);
-        assert_eq!(caret_from_x(value, 999.0, measure), value.len());
-    }
-
-    #[test]
-    fn scroll_x_updates_only_when_caret_leaves_viewport() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        // 100 chars, 10px each.
-        let value = "x".repeat(100);
-        store.set(id, value);
-        store.focus(id);
-
-        let available_w = 50.0;
-        let text_w = 1000.0;
-
-        // Caret at start: no scroll.
-        store.set_caret(id, 0, false);
-        store.update_scroll_for_caret(id, 0.0, text_w, available_w);
-        let (_v, _caret, _sel, scroll, _scroll_y) = store.get_state(id).unwrap();
-        assert_eq!(scroll, 0.0);
-
-        // Jump caret to end: scroll to max.
-        store.set_caret(id, 100, false);
-        store.update_scroll_for_caret(id, 1000.0, text_w, available_w);
-        let (_v, _caret, _sel, scroll, _scroll_y) = store.get_state(id).unwrap();
-        assert_eq!(scroll, 950.0);
-
-        // Move caret slightly left but still within viewport: no scroll change.
-        store.set_caret(id, 99, false);
-        store.update_scroll_for_caret(id, 990.0, text_w, available_w);
-        let (_v, _caret, _sel, scroll, _scroll_y) = store.get_state(id).unwrap();
-        assert_eq!(scroll, 950.0);
-    }
-
-    #[test]
-    fn checked_mutators_return_changed() {
-        let mut store = InputValueStore::new();
-        let id = Id(1);
-
-        // Starts unchecked by default; setting to false is not a change.
-        assert!(!store.set_checked(id, false));
-        assert!(!store.is_checked(id));
-
-        assert!(store.set_checked(id, true));
-        assert!(store.is_checked(id));
-
-        // Setting the same value is not a change.
-        assert!(!store.set_checked(id, true));
-        assert!(store.is_checked(id));
-
-        // Toggling always changes current state.
-        assert!(store.toggle_checked(id));
-        assert!(!store.is_checked(id));
-    }
+/// Convert `html::Id` to `input_core::InputId`.
+#[inline]
+fn to_input_id(id: Id) -> InputId {
+    InputId::from_raw(id.0 as u64)
 }
